Refactor context management: Replace LogBookContext with dedicated contexts for Auth, Blog, Chat, Playlist, UI, and UserData. Update components to utilize new contexts for improved modularity and maintainability.
