+++ src/components/post/PostEditor.jsx
@@ -1,12 +1,12 @@
 import { useNavigate } from 'react-router-dom';
 import { useState, useRef, useEffect } from 'react';
-import { useLogBook } from '../../context/LogBookContext';
+import { useBlog } from '../../context';
 import PostToolBar from './PostToolbar';
 import PostEditorModal from './PostEditorModal';
 
 const PostEditor = ({}) => {
     // useContext
-    const { markdown, setMarkdown, postTitle, setPostTitle } = useLogBook();
+    const { markdown, setMarkdown, postTitle, setPostTitle } = useBlog();
 
     // useNavigate
     const navigate = useNavigate();
