+++ src/context/index.js
@@ -1,0 +1,8 @@
+// Context exports
+export { AuthProvider, useAuth } from './AuthContext';
+export { ChatProvider, useChat } from './ChatContext';
+export { BlogProvider, useBlog } from './BlogContext';
+export { PlaylistProvider, usePlaylist } from './PlaylistContext';
+export { YTPopupProvider, useYTPopup } from './YTPopupContext';
+export { UserDataProvider, useUserData } from './UserDataContext';
+export { UIProvider, useUI } from './UIContext';
